reestructured back: inactivePreference as an array, refactor components into different files from FilterArea, handle Delete
